Add blank lines to shorter bookmarklets
Keeping all the code on the same line allows desktop users to triple-click to select. The blank lines give mobile users another line to include in their selection, which makes selection a lot easier than having to scroll horizontally bit by bit and expanding the selection in the same way.